llsf_msgs: Bugfix,AchieveWorkpiece,Gate int32>int64
This caused crashing when trying to deliver a wp by a slave robot.

Cause, trying to incorrectly fill the delivery mps_instruction with a
wrong gate value.
The optional gate field in AchieveWorkpiece msg, was never filled
correctly. The type of the term that contains the gate number was
not a int32. (even though i expected it to be so. Not sure why.Check
later)