gazsim-scripts: fix calculation of per-team robot number
The calculation of the number of robots per team was wrong if setting
the first robot with -f, e.g., "-n3 -f4" to play with only magenta
robots.

This is fixed with this commit. First, it needs to be done after
parameter processing to give the same results independent of the order
of arguments, e.g., "-n3 -f4" vs. "-f4 -n3". Second, the calculation of
number per team did not take the "first number" argument "-f" into
account.