openprs-agent: PlayGame BASIC, Produce C0,C1.
This is the huge commit that includes many changes commited for the sake
of reference and as a milestone. Those changes will be broken down into
several commits on the branch mgomaa/openprs-agent. This commit should
not be used for further development.

Changes: Producing C0 and a C1 after each other on the start of the
production phase using 3 robots. There is not Task priority or smart
allocation. All Tasks related to a chosen goal, are intended in parallel
on their own separate intention, as a root of the intention graph.

Each Task Sleeps on its preconditions till. When preconditions are
satisfied, A free robot is chosen to achieve the task.
Each Task is constituted of a Sequence of Achieve-Workpiece Goals that
will be sent to the chosen allocated agent, and wait conditions to make
sure the task wont be continued till the expected resulting workpiece is
is asserted to the master's database. A task is most often a workpiece
gripping achievement, followed by an WP insertion Achievement. The
Achieve-workpiece goals blocks and starts a monitor that tracks the
achievement status of this workpiece. Only when an Achievement Succeeds
or Fails can the Task execution be continued.

All Machine handling and instructions are left to the individual agents.
Achieve-workpiece goals could have an mps-name as an achiever. This
models that an mps will be responsible for a workpiece transaction
from a place to a different place (from its input to its output for
instance). The master broadcasts this Achieve-workpiece to all agents.
The agent responsible for the handling of this workpiece can then use
this information to figure what to instruct the machine when needed.
The status of the Achievement is published globally.

All worker agents are treated equally by the Master-of REASONING.
Agents have not idea about the goal or even the task they are executing.
The Master-of REASONING send Achieve-Workpiece

For example, The TASK-Cx--MOUNT-A-RING does the following sequence:
	-Detect ,based on the current state of procution of a WP, which
	 Ring will be mounted next.
	-Waits till there is enough payment material, at that RS slide.
	-Waits for a agent to IDLE and allocate/book it for the task.
	-Request an Achieve for a gripping action at the agent (where ever
	 the workpiece is, left for the agent)
	-Double check if the Achievement resulted in the correct WP
	-Request 2 Achievements in parallel with 2 different achievers
	  one for the agent gripping the Current Wp to insert it at MPS.
	  one for the MPS with the 'expected' WP at its output [Dumy].
	  (Both goals will block till reaching a final status)
	-Double Check for the expected workpiece at RS output
	-Release Agent
	-Intend Again in case more rings needed