openprs-agent: Co,Achieve-Monitor remember status
PROBLEM was: Sometimes the Achievement is finalized too quick and deleted
at some point during the execution of this op. If that happens then
The state will not be tracked correctly resulting that either a wait
condition to never be satisfied. Or A succeeding the Goal no matter
What happens.
.To fix this u need to keep track of the status a bit differently
Remembering what cozed the thread to wake