openprs-agent:keep multiple lock-in-use facts
allow to keep multiple lock-in-use facts for multiple machines.
lock-in-use fact keeps track of where the machine is. It was an ff 0
which means I can only keep one unique instance per robot. This caused
locks to not get released if the fact was granting another lock.
This fact is crucial in deciding to release the lock.