asp-planner: Fix warning with GCC7
Ensure the length is calculated at compile time and make it big enough
for a short. So reduce the value of the calculation to a short. Because
there can be no longer idle time than 900 seconds per robot (as of 2016)
a short is still big enough to hold the value.

error: ‘ seconds’ directive writing 8 bytes into a region of size between 4 and 13 [-Werror=format-overflow=]
 AspPlanerThread::loop(void)
 ^~~~~~~~~~~~~~~
note: ‘sprintf’ output between 17 and 26 bytes into a destination of size 21
       std::sprintf(idleString, idleFormat, idle);