rcll2016-smt: Fixed task sync Id [Task Sync hack]
Fixing the task sync id will ensure that there is ever one running task
between all agents.

This is done to avoid a bug in the worldmodel-synchronization that
caused a slave agent to sometimes have 2 tasks.

This happens because the retraction of facts performed by the Master
are never really communicated to the slaves (the master only published
what he has in the its worldmodel but Not what he removed).
So race-condition occurs if a task was finished on the slave and
retracted. Then asserted again because it was communicated with the
master's a little outdated worldmodel.