Clips-smt: Add action IDs for concurrent execution
For every action we add to protobuf a counter is increased.
This counter is saved into the action via aciotn->set_id(ctr).
After adding the final action of a MACRO-action the corresponding id is
saved separatly into action_id_last.
The parent_id of an action indicates dependencies and is filled by the
value in action_id_last.