Clips-smt: Refactor assignment of team and bugfixes
Omit two blocks of code with teams CYAN and MAGENTA hardcoded.
Now extract team from protobuf into an variable and construct names of
mps.

Fix the bug that mps in one skill was not set.

Fix the bug that parent_id was added to last skill of MACRO-action
instead to the first skill.