Clips-smt: Add all required parent-ids
Version before assumed that the order of skills is used.
New version assigns parent-ids to all skills if they are needed.
Now the skills are related as follows:
	- The skills inside a MACRO-action are perforemd sequentially as
they get the parent-ids of the skill before
	- The dependencies of the MACRO-actions are displayed as
parent-ids mapping to the id of the last skill of an MACRO-action
	- Some skills do not require parent-ids as they could be
executed directly (here one should execute in the order of increasing
ids)