Clips-smt: Handle usage-order of robots dynamically
Before we added constraints to enforce that first R-1 is used, then R-2
and finally R-3 in order to omit traffic jams at the beginning of the
execution.
Now we do not restrict the usage-order in our encoding anymore and add a
permutation such that the first used robot (no matter which one) is
added as R-1, the second as R-2 and the third as R-3.