Clips-smt: Add parent_id to prevent occupying of cap station
We add the id of the last skill in the 'discard cap
carrier'-MACRO-action sequence as a parent_id of the move skill in order
to obtain the final c0 product.
With this dependency between the skills we prevent that a robot, which is
assigned to ONLY take the final product to the delivery station, will
not block the 'discard cap carrier'-MACRO-action sequence.