protoboard: Templated plugin to convert protobuf to blackboard
This plugin still resides in the domain repository, but all
domain-independent code is separated. Domain-specific
protobuf-blackboard mappings are specified as template specializations.
So in effect, this is not a plugin, but a plugin template wich can be
moved to the core repo after the source layout has been refactored a
bit.