protoboard: fix pb->bb conversion for repeated fields
So far, only a direct mapping from one or more incoming protobuf
messages to one or more blackboard interfaces could be specified. This
commit adds the ability to extract a repeated field from a message and
map each element to a specific blackboard interface. Blackboard
interfaces are opened dynamically for each repeated field, and closed
again for each field that no longer exists (in relation to the previous
message).