Clips-smt: Add constraints for c2-c3 complexity
We added the necessary actions for c2-c3 complexity and take into
account that certain ring colors require a payment before mounting.

We added the goal_id to the skills sent via protobuf.
Note that the goal_id is 0 as we only cover one order which is the
first at this moment.

Some refactor, as we do not need the state1 anymore using MACRO actions.
The global variables to take care of the different complexities have
been renamed.