protoboard: implement matching for repeated fields
In the case of RCLL Orders, each repeated Order field in the OrderInfo
message has a persisting identity specified by its order ID. Each order
ID is either new, or it needs to be mapped to the specific blackboard
interface holding that order. The appropriate blackboard interface
should then only be changed if some of the other fields have changed.

This commit changes the pb->bb conversion templates to require a method
corresponds_to() that must return true if a certain pb message has the
same identity as the respective blackboard interface.