Clips-smt: Generalizes constraint about robot order
Before we had a two component version of the robot order.
Plan step 2 is encoding that only R-1 or R-2 might be used
and from plan step 3 on we state that R-3 is only used if R-2 has been
used before.
Now we take only the generalized version but from plan step 2 on.

The idea behind the constraint encoding the robot order is to only look
to one equivalence class of robot orders.
Because some robot assignments are identical with respect to a renaming
of robots.