velocity_share: share positions and velocities via robot-memory
This plugin shares position and velocity information among robots via
robot-memory.
Each robot is identified by its "robot-name", (taken from the configuration)
and creates an object for itself if not available in the "memory.robots" collection.

The position is gathered from the Position3DInterface "pose".
The velocity is gathered from the MotorInterface "motor_if".