clips-exec/rcll2017: Evaluate Produce-C0, order-wp relation
Upon Evaluation of the completed goal 'PRODUCE-C0', assert the
relation between the WP dispensed by the BS, and the order that
it was intended for.

This will be used later to relate workpieces to the corresponding order
they were meant for. Will be useful for tracking the progress of a WP
and selecting further processing goals on that WP (Ex, DELIVERY or
MOUNT-RING).