clips-exec/rcll2017: formulate goal 'DELIVER'
Deliver workpieces related to an order when they have a cap
and ready at the CS output side. Deliver only for orders that
still require delivery.

At this stage, we have no consideration of time windows of delivery.
TODO: add all the time related conditions to the LHS of the goal
formulation rule