clips-exec/rcll2017: fix, Guard common Eval goal vs Produce-C0
PRODUCE-C0 goals are evaluated from its own goal evaluation rule.
Yet the common Evaluate goal rule used to fire in place.

This was caused by the faulty assumption that the order of the rules in
a file could be taken as an indicator to what fires first. It is clear
now that this assumption is a Myth. Rules with the same salience fire
arbitrarily. Guards are needed