clips-exec/rcll2017: introduce PRODUCTION-MAINTAIN goal
the production goals are now part of a maintenance goal
PRODUCTION-MAINTAIN. The goal is created on the appropriate
refbox status and when domain is loaded.

This allows to have a better grouping of the production goals and
to be able to easily distinguish their custom behavior through
the stages of their goal life cycle.