cfg laser-filter: unify front and back laser filters
Do not assume the front laser to be a 1080 laser and the back laser to
be a 360 laser. Instead, provide filter chains for both possibilities,
which can be adapted by enabling/disabling the right filters.

In particular, for a 1080 laser, enable the respective downsample
filter. For a 360 laser, use the 360to360 laser instead. This way, we
can always use the '$name 360' laser topic for further processing.

This always easier host overrides for robots with a 1080 laser in the
back.

On the way, also rename the filtered front laser name from filtered-1080
to front-filtered-1080 so we always have the original source in the
name.