conveyor_pose: transform model cloud gripper->cam
This should be an easy way to transform the model cloud so the
correspondence grouping returns a helpful transform. Just position the
gripper at the desired origin of the conveyor, and then transforming it
will put the model at the origin of the realsense camera.

Also the model file is written atomically now.