goto: check failing tf and wait until tf becomes available
We cannot always expect tf to return a transform. Instead, if tf returns
nil, wait for the transform to become available before going into the
INIT state. This means that the skill does not immediately fail if the
transform is not available. Note that we still fail immediately if the
navigator is not running.

This fixes a bug where lua complained about 'cur_pose' being nil while
executing enter_field.