cx/rcll2017: add locked predicate and locking actions
A lock can be on any object of the domain. The action 'lock' locks the
object, i.e., it does not finish until the lock has been acquired.
The two actions 'unlock' and 'eventually-unlock' are the same in the
domain model; the second can be used in the execution model to express
that the lock will eventually but not necessarily immediately be
released. This can be used for distance-based unlocking.

In the execution model, 'locked' should be a sensed predicate. The
actions 'lock' and 'unlock' should be waiting for sensed effects, the
action 'eventually-unlock' should not.