mps-laser-gen: setup mps box rendering relative to map
The box_filter of the laser-filter plugin allows to filter oriented
rectangles from a laser. We make use of this to provide a laser that
does not include mpses. This can later be used for velocity changes
based on the fact that we only get dynamic obstacles relative to the
robot.