cx/rcll2017: add location-lock actions with distance-based unlocking
The actions location-lock and location-unlock lock one side of an MPS.
After location-lock succeeded, the predicate
(location-locked ?mps ?side) holds. The predicate is a sensed predicate,
which allows us to unlock the location lock without waiting for the
effect to occur. If an unlock is requested, the lock will eventually be
released if the robot reaches a certain distance threshold to the
location.