conveyor_pose: properly catch transform errors
With the new gripper, we now have a dynamic transform in the path from
base_laser to cam_conveyor, so there are new race conditions when
transforming along that path. Properly catch those so we don't crash.
Also, retry transforming laser-line poses with Time(0,0) if transforming
at the line's timestamp fails due to an ExtrapolationException. This
removes any logging since we silently work around the race, so the
actual cause of the race still warrants investigation.