arduino: setup new ArduinoInterface for 4 axis feature
The new ArduinoInterface contains several changes towards the new 4 axis
to be driven by the Arduino.
The MoveXYZAbs message supports setting goal positions relative to a
proper frame. This can be a frame that is fixed relative to the robot
(to move to absolute positions) or fixed relative to the gripper (to
move to relative positions - relative to the gripper).

The MoveXYZRel message supports motion just relative to the gripper
without providing a frame. However, in the future we might merge both of
the new messages to a "MoveXYZ" message.

Apart of the new messages we remove those only needed by the former Z
positioning and include others for convenience ("ToHome" and
"Calibrate"). The "Calibrate" message is supposed to let the gripper
re-initialize the currently known position by moving the different axes
backwards until we reach an end-stop.

The "ToHome" message moves the 3 axes to X and Z being 0 and Y being in
the middle of the axis.