static_transforms: setup static transforms for 3 axis gripper
The new gripper requires a set of new static transforms to be set up.
This includes a gripper_home frame defining the home position of the
gripper. This is the position of the new gripper after it reached its
home position.

The conveyor_calib frame is supposed to be used as a goal frame to
record new models of the MPS conveyor.