cx/rcll2017: added a goal to produce the easiest possible c3 product (independent from any orders) to test the simulation.
The goal GARBAGE just picks a red base and then places 3 rings on it that all need no additional bases.
There are always exactly 2 ring colors that need no additional bases and they are mounted by different ring stations.
The goal always mounts the different ring colors in alternating fashion.
It is for testing the silumation and does not serve any other purpose. It only works if no additional bases
were fed to the ring station before and the product will get discarded after the 3rd ring is mounted.