cx/rcll2017: remove expired cx products from ring stations
When cx products expire it often is still viable  to produce them as the
penality for late deliveries is low. Being able to clear ring stations
is still important to prevent the production to running into
potential deadlocks. So for now this method follows the same decision
procedure as it is in place for C0 products.