cx/rcll2017: improve goal lifecycle of parent goals
Always let the child goal go to the next mode before switching the
parent to the next mode. That means the parent is only committed if a
child was committed, and it is only dispatched if a child was
dispatched. Also make sure that the parent already reached the right
mode, e.g., only dispatch a child goal if the parent has been committed.