motor_move: switch from timeout to navigator:is_final()
The colli seemed to have some issues regarding a faulty information
about being final or not. This lead to situations where two
components tried to adjust for a certain position, the colli and
motor_move. To overcome this issue, motor_move waited for one second,
giving the colli a bit more time to get "really" final.
However, this issue does not occur with our current setup of move_base.
Thus we can simply transit to INIT when the navigator interface is
final.