drive_to_machine_point.lua: new skill for mps approach
The new skill drive_to_machine_point will take a certain navgraph pose
(e.g. C-CS1-I) and try to detect a laser line matching the orientation
of the navgraph point while approaching this navgraph point.
Once the laser line exceeds certain tunable thresholds it is considered
as the laser line of interest and the drive_to_local is interrupted
followed by a set of motor_move executions to approach the MPS.
The skill allows to approach preconfigured Y-positions in front of the
MPS (e.g. "CONVEYOR").