cx/rcll2017: report opponents mps to navgraph aswell
Per request of nlimpert: To also have access to the coordinates of the opponents mps
(e.g. to also register them as static objects) the mps data from the refbox is used
to also compute the mirrored locations and orientations and add them as found-tags.

This also fixes the generation of the waiting positions.