cx/rcll-smt-c0-robust: Send mps 'from' of move action via plugin
Before we kept track of the 'from' information for the move action on
the clips side. Now in order to be more robust in case we call the
plugin again we use the locations generated by the plan for the final
plan actions.

Translate START-I as the node name team-ins-in for the distance
computation using the navgraph.