cx/rcll2018: move retraction of rs-fill-priority facts
Previously rs-fill-priority facts got removed on the formulation of
fill-rs like goals (either FILL-RS ord FILL-RS-FROM-BS).
Those two goals were exclusive, in preparation of a new
FILL-RS-FROM-SHELF goal which can get formulated along with
FILL-RS-FROM-BS this exclusiveness is no longer given and formulating
one of the goals would already remove the rs-fill-priority facts that
the other goal would need.

To account for that as well as to have a more clean way of treating
rs-fill-priority facts they now get removed whenever the production
maintenance goal is evaluated (so a when a production subgoal finished
and thus possibly changed the rs priorities during its execution).