cx/rcll-smt: Fix bugs in plugin
We did not follow which capcarrier wp is where.
Thus we did adress the wrong CCBx or CCGx for a wp-get.
As a fix we changed the planning behavior that for feeding
additional bases we always get a base at the base station.

Adapt the correct ids for ring colors (orange and yellow was swapped).