cx/rcll2018: remove resource lock on the delivery station
Since the state of the delivery station is not changed after a delivery
there is no need to lock it as resource on goal level.
Without a resoource lock on the ds multiple products can be delivered,
while letting robots wait for the ds to get free again when another
robot uses it.