cx/rcll2018: the production of exclusive products requires token
Whenever the production of a product that has a exclusive complexity
is started, the PRODUCE-EXCLUSIVE-COMPLEXITY token needs to be locked
as a resource. Since resource locks get released after the corresponding
goal is evaluated this ensures that the wp-for-order fact gets created
before the production of another exclusive complexity product can be
selected. After that wp-for-order fact is present no other exclusive
complexity product can be started and so ensures that the exclusiveness
holds after the token is released.