cx/rcll2018: remove resource-locks on mps that just get cleared
Whenever a goal only uses an mps to get a wp from the output
there is no need to have it as resource lock, because the functionality
of the mps does not change. Plans of other agents that are formulated
before an mps is cleared can also be formulated after the wp is taken
out of the machine.