cx/rcll2018: Exclude RESET-MPS on unexpected input (wrong assumption)
One robot's absence of knowledg about if that wp has a mutex.
does not impley that no one other robot has that mutex.
(ie, it could easly be the case that R-1 has the WP mutex, yet
R-2 doesn not try to get it, hence does not know about it,
..R-2 resets the machine)