cx/rcll-smt: Adapt value types to use of rob-mem-sync
Before we converted and used some values as symbols. But using
rob-mem-sync converts them back to strings. So we have to adapt all
values to strings such that there is no difference for rules used by
R-1 (currently the fix leader of the group which plans) and the other
robots.