cx/rcll-smt: Adapt value types to use of rob-mem-sync
Besides adapting the value types for making the rules uniform to all
robots we use now a new workflow for deciding if a plan-action matches
all its parents-ids. As there happen some conversions on the
parents-ids via the rob-mem-sync in the sync of wm-fact dep, now only
R-1 (who is generating these facts) is processing this wm-fact. If a
plan-action matches its parents-ids then a wm-fact dep-match is
asserted and synced to the others.