plexil2016: read game state and game phase from refbox msgs
Read the current game from the refbox and set them as global so they can
be looked up from within a plan. Modify the SkipCondition, so we don't
skip refbox messages anymore just because the team is set. Instead, move
the team setting into a separate block and only skip that block.