buildkite: add pipeline configuration based on Fawkes Core
Use a dynamic pipeline that currently uses the core pipeline with the
necessary modifications. This allows us to maintain the buildkite
configuration in a single place for the time being. Add a basic example
how this could be extended.

We may later opt to have a separate pipeline configuration if that
becomes necessary or is desirable.