Revert "cx/rcll2018: prevent prioritized prefilling of rs when not needed"
This reverts commit 7c56c4565ed64e68f8231a3ac743ec664df6ebcf.
The numbers of rings are of type SYMBOL and written out as they are part of
the PDDL domain. They can not be compared via '<' but it is not needed
here anyways. The case that ?rs-before is strictly greater than
?ring-num is implicitly prevented by the following precondition on rs-sub.
rs-sub models subtraction on the domain of THREE, TWO, ONE, ZERO.
In the case of ?rs-before > ?ring-num there is no rs-sub fact where
?ring-num is the minuend and ?rs-before is the subtrahend as the result
would be negative.