arduino: add add_setting function
Grbl has a lot of settings to set. As the syntax to set them is
different from the rest of the commands, a different method is
introduced.

The setting_map includes an enumeration identifier for each setting.
Further the type of the setting is fixed, to simplify the transformation
into the correct string.