gazsim-scripts: do not use deprecated -e/--command
gnome-terminal does not support the command option anymore. Instead, we
need to create a new window first, and then pass a command to that new
window that opens up all the new tabs. This requires one additional bash
shell that creates all the tabs, as we would otherwise need to call
multiple gnome-terminal commands from within the gnome-terminal command,
which is not supported by the new gnome-terminal options.