gazsim-scripts: add screen support
The simulation can now also be started in a screen session called
'gazsim'. The screen session will be detached and remain detached. You
can manually attach to it with 'screen -x -S gazsim'.

Start the initial screen session with /usr/bin/sleep 1 just to make sure
that the other commands are run before the session is terminated.