gazsim-scripts: add informative message for first tab in terminal
After #46 got merged, there is one tab which has the only purpose of
spawning the other terminals. Print an informative message. This is
relevant for the case that "Hold the terminal open" is enabled for
debugging, as not to confuse anyone (the first tab used to be the
Gazebo output).