arduino: move definition of setting constructor into header file
For correct compilation, the definition of a template member function must
be either ..

..inside the header:
The function is then compiled for every translation unit, for every used
type.

..inside a separate cpp file:
The function cannot be compiled then whenever it is called, since its
code is not available. Using explicit instantination, the compiler is
hinted to prepare versions of the function for given argument data
types.

To be consistent, i.e. have all function definitions at one place, the
second option was chosen. Unfortunately, doxygen is not able to handle
explicit instantination of template member functions, iff the function
is a constructor. Consequently, this commit changes to option 1.