arduino: Use poll instead of run_one
The io_service automatically stops whenever no tasks are in the queue
anymore. Afterwards, the io_service needs to be restarted using the
reset function.

To prevent the io_service from getting workless, the run_one function
was used, together with a deadline_ timer whose timeout was set to
infinity if not used.

This commit introduces the usage of an work object, which forces the
io_service to not stop until the complete thread is destroyed.

Further, all handlers are executed when they are ready, which is
accomplished by using the poll function.