plugins: rename old conveyor_pose to conveyor_plane
Preparation for extending the conveyor detection pipeline to:

laser-line -> conveyor_plane -> conveyor_pose,

i.e. the old (plane detecting) conveyor detection would be used as an
initial estimate for the new (ICP-based) conveyor detection.

This is expected to increase the reliability of the ICP significantly,
since bad initial estimates from using the laser lines directly are the
primary source of errors.