cx: add simple goal sub-type
Simple goals do not have sub-goals. They are suitable to act as leaves
in a goal tree or as stand-alone goals.

The goal lifecycle is automated such that a user only has to formulate,
expand and evaluate the goal. Goal selection and cleanup is automated for
simple goals that have no parent.