cx: add run-endless goal sub-type
A run-endless goal is a maintain goal that gets reformulated after it
is evaluated and all sub-goals are cleaned up and a minimum time elapsed
between the last formulation. This formulation frequency can be
configured via the parameter frequency. Therefore run-endless goals are
suitable to act as parent goals for continuous or periodic sub-goals.

Goal selection,expansion and evaluation is up to the user, the remaining goal
lifecycle is handled automatically. A run-endless goal needs to have at
least one sub-goal before it is expanded. Then it selects the sub-goal with the
highest priority.