cx: adapt production goal creation to goal sub-types
To begin the migration to goal trees assign a sub-type to each
production goal (SIMPLE for the ACHIEVE goals and RUN-ENDLESS for
MAINTAIN goals). Also rename goals that previously had ACHIEVE or
MAINTAIN in the class name to be more consistent.
Goal meta facts become obsolete as well because of the meta field in
goals, therefore they are removed.