cx: move globals of max tries into goal reasoner
The goal reasoner is loaded before execution monitoring and goal
production as it defines the saliences of the goal lifecycle. Globals
that define the maximum number of retries for goals and actions might be
used within all three files. So it makes more sense to define them in
the goal reasoner.

This change is mandatory when moving goal evaluation rules to the goal
reasoner.