arduino: Update Makefile for arduino sketch
The arduino is linked to different file handles, depending on the
OS and specific setting.
For the robotino laptop, an udev rule links the arduino automatically
to /dev/arduino.
The new Makefile version accepts an optional variable definition of
DEV_ARDUINO, to support the variety of file handles for the arduino.