arduino: Correctly compute time
As the time is stored in a 16 bit register, but the data bus is only
8 bits wide, reading the time without risk is not possible easily,
without additional safety mechanisms.

The arduino offers a safety mechanism, by copying the High byte into
a shadow register, whenever the low byte is read.

This commit changes the order in which low and high byte are read,
fixing timing errors.