cx/rcll2018: fix location unlock handling
React to location unlock response and delete location-locked facts after
locking as they are not handled by the domain-effects due to distance
based unlocking.

When unlocking a location the corresponding plan action is forwarded to succeed
without actually applying the effect. This is due to the fact that the
lock is only freed when the robot actually moves away from the location
afterwards.

Co-authored-by: Till Hofmann <hofmann@kbsg.rwth-aachen.de>