cx: make run-endless goal to only take exactly one sub-goal
Maintain goals are intended to maintain a condition by executing one
sub-goal whenever said condition is violated. The old behavior was
partially duplicating the behaviour of run-one goals which is not
desired.

For the production goal tree this means that the maintain goal gets a
PRODUCTION-SELECTOR sub-goal of sub-type RUN-ONE-OF-SUBGOALS that picks the
currently best possible step to fulfill the orders.

Co-authored-by: Till Hofmann <hofmann@kbsg.rwth-aachen.de>