arduino: Add verify and safe_flash targets
When preparing the robotinos for the competition it is useful to make
sure that all arduino are flashed with the newest firmware. One way to
do this is flashing them every time when the makefile is run.

However, the amount of write operations for the arduino's flash is
limited. A better approach is to verify the firmware on the arduino
first and only flash the firmware if the verification is unsuccessful.

The target verify only verifies the firmware, while safe_flash also
flashes the firmware, if the verification was unsuccessful.