config: setup ethernet access for lasers
The TiM lasers are not supposed to be used via USB permanently but
rather for debugging and configuration purposes. The actual usage of
these should happen via ethernet.
This commit changes the use of the two lasers via ethernet and sets up
IP addresses that are to be configured on the lasers with the SOPAS
Engineering tool.