goto: use navgraph {ori, target}_tolerance
With the changes introduced in f0897ab0d in the Fawkes core repository
we can now set a certain tolerance instead of using the one configured
in the ros-navigator.
This allows to use the tolerances of navgraph points in
particular.
In case either the ori or target tolerance of the navgraph point are
missing, we send a message without a tolerance.