cx: wait for rs that is down, when it can be filled
Allow to formulate goals to fill a rs when the machine is currently
down.
If one rs is full, the other one is down and a bot holds a workpiece to feed,
then previously the workpiece had to be discarded.
If there is no more valuable task to do, then the follow up goal would
likely be to get a workpiece to feed, which would block other
machines in the process. Therefore waiting instead is the better choice
in that scenario.

Co-authored-by: Till Hofmann <hofmann@kbsg.rwth-aachen.de>
Co-authored-by: Daniel Habering <daniel.habering@rwth-aachen.de>