arduino: Make A axis microstepping configurable
By changing the steps the A motor has to do to open/close the gripper
the user can decide which microstepping mode he wants to use.

As setting the exact right number can be sometimes a little bit tricky,
instead of giving the microstepping mode the user has to provide the
exact step number to open/close the gripper. However, a handy tip to
compute this number is given in the config file.