tag_vision: set pruning threshold to -100
The visibility history threshold for pruning unseen tags from the
tag-vision/info interface was set to -1000, which takes way too long to
reach. When the bot moved around and saw lots of tags, it was cluttering
the interface up with unseen tags, leaving no room for new, actually
visible tags.