goto: fix place to zone regex
When the place was assumed to be a zone, the skill did not properly
check whether the zone might actually be a navgraph place.
This is fixed by making sure the regex only matches if the setup place
only starts and ends with a proper zone name (e.g. M-Z21).
Otherwise a place named "WAIT-C-Z23" would match the regex leading to a
skill failure.