cx: add a rs-ring-spec property for RING_NONE
rs-ring-spec was previously only defined for the available ring colors,
making it difficult to determine ring specs of an order independent from
its complexity. To mitigate this RING_NONE gets the ring-spec NONE,
making it distinguishable from the numbers ZERO, ONE, TWO and THREE as
well as from the spec NA which is used to indicate that the info from
the refbox is not available yet.