cfg: Make host_$host.yaml normal include file
The host-specific tag in the config.yaml file is used to describe the
location where updates of the config are written by the fawkes
framework. Consequently, it potentially changes during every run of
fawkes and is thus not suitable for pushing to the repo.

This is why the host_$host.yaml is now degraded to be a normal config
file. Its host specifity is now purely given by its name.