cx: remove mps-handling from production goals
Add a new maintain goal that handles preparing and modelling
the mps processes. If a goal needed to prepare a mps as the last thing
to do, it now asserts two facts, stating the type of prepare
action and the type of processing action. As soon as the
mps is IDLE and there is no workpiece at the output, the new maintain
goal takes care of sending the prepare message and modelling
the effects of the mps process