gazsim-scripts: do not load default plugins if meta plugin was specified
If a meta plugin has been specified on the command line with `-m
plugin`, do not load any other plugins by default. Previously, the
specified plugin was loaded in addition to the default simulation
plugins. However, if we want to simulate without gazebo, this is no
longer desired.

This means that if you want to load a different meta plugin in addition
to the gazebo plugins, you need to specify a meta plugin that includes
all those plugins.