cx: replace value NONE of ring-color by ZERO
Value NONE was intended as a valid domain number to specify rs-ring-spec of
RING_NONE in order to make writing of rules easier that react upon
workpieces of different complexities. The value NONE could be
distinguished from ZERO (no additional base needed) and NA (no
information available yet). It turns out that we also need to use the
number value of rs-ring-spec as argument in rs-sub. It seems like the
fine distinction between, whether a ring of color RING_NONE has a
rs-ring-spec value of NONE versus a value of ZERO is a small gain
compared to the workaround of creating artificial mathematical operations like
TWO - NONE = TWO.
Therefore value NONE is getting replaced by ZERO.