cx: distinguish between order and workpiece related meta facts
It is possible to have multiple workpieces for the same order, because
products can be ordered multiple times (quantity-requested >=1). To
distinguish between the progress of different workpieces, progress
related information is now saved as wp meta facts. This includes:
 - points-current
 - bases-missing
 - rings-missing
 - estimated-points-next-step
 - estimated-time-next-step
 - estimated-time-total
This also means, that the above information is not available when
deciding whether to start an order or not (because there is no
associated workpiece yet).