cx: rename order meta points-total to points-max
The value of the points-total fact is intended as reference to compare
against estimated point gains that consider temporal information.
Therefore it makes sense to save the theoretical maximum of points that
each order can yield. The meta facts for estimated point outcomes are
named 'estimated-points-total'. To hopefully avoid confusion between
those facts and the points-total facts we rename points-total into
points-max.