cx: refactor wp meta facts
It is unecessary to have facts for the estimated point gain and time
required when the next production step finishes. The information is
already present in one of the lists from the order meta facts
points-steps and estimated-time-steps. Instead we can keep track of the
next production step that has to be performed in order to progress a
product.