skills/drive_to_machine_point: wait before mps_align
In case the robot ends its motion with a rather large rotation in front
of the mps the laserline of the mps of interest first has to settle for
some time.
As a result we can't rely on the laserline to be at the appropriate
position unless we gave it some time to settle.